Jarvis3/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.4
google-trans-new==1.1.9
googlesearch-python==1.2.4
idna==3.7
joblib==1.4.2
nltk==3.8.1
numpy==2.0.0
playsound==1.3.0
pocketsphinx==5.0.3
PyAudio==0.2.14
pycparser==2.22
pygame==2.6.0
pypiwin32==223
pyttsx3==2.90
pywin32==306
regex==2024.5.15
requests==2.32.3
six==1.16.0
sounddevice==0.4.7
soupsieve==2.5
SpeechRecognition==3.10.4
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
wikipedia==1.4.0